Move echo in its own file.
