+++ inst/include/dplyr/hybrid/hybrid.h
@@ -22,16 +22,6 @@
 namespace dplyr {
 namespace hybrid {
 
-inline SEXP echo(SEXP x, const Summary&) {
-  return R_UnboundValue;
-}
-inline SEXP echo(SEXP x, const Window&) {
-  return x;
-}
-inline SEXP echo(SEXP x, const Match&) {
-  return Rf_mkString("echo");
-}
-
 template <typename SlicedTibble, typename Operation>
 SEXP hybrid_do(SEXP expr, const SlicedTibble& data, const DataMask<SlicedTibble>& mask, SEXP env, const Operation& op) {
   if (TYPEOF(expr) != LANGSXP) return R_UnboundValue;
@@ -329,12 +319,8 @@ SEXP hybrid_do(SEXP expr, const SlicedTibble& data, const DataMask<SlicedTibble>
       Column x;
       int n;
 
-      if (expression.is_unnamed(0) && expression.is_column(0, x) && expression.is_named(1, s_n) && expression.is_scalar_int(1, n)) {
-        if (n > 0) {
-          return lead_1(data, x, n, op);
-        } else if (n == 0) {
-          return echo(x.data, op);
-        }
+      if (expression.is_unnamed(0) && expression.is_column(0, x) && expression.is_named(1, s_n) && expression.is_scalar_int(1, n) && n >= 0) {
+        return lead_1(data, x, n, op);
       }
 
     } else if (expression.is_fun(s_lag, s_dplyr, ns_dplyr)) {
@@ -343,12 +329,8 @@ SEXP hybrid_do(SEXP expr, const SlicedTibble& data, const DataMask<SlicedTibble>
       Column x;
       int n;
 
-      if (expression.is_unnamed(0) && expression.is_column(0, x) && expression.is_named(1, s_n) && expression.is_scalar_int(1, n)) {
-        if (n > 0) {
-          return lag_1(data, x, n, op);
-        } else {
-          return echo(x.data, op);
-        }
+      if (expression.is_unnamed(0) && expression.is_column(0, x) && expression.is_named(1, s_n) && expression.is_scalar_int(1, n) && n >= 0) {
+        return lag_1(data, x, n, op);
       }
 
     } else if (expression.is_fun(s_in, s_base, ns_base)) {
